Successfully Integrate 4 pipelines
