+++ dvc.lock
@@ -136,3 +136,5 @@ stages:
     cmd: dvc repro pipelines/fake_classification/dvc.yaml
   emotion_classification_pipeline:
     cmd: dvc repro pipelines/emotion_classification/dvc.yaml
+  age_gender_pipeline:
+    cmd: dvc repro pipelines/age_gender/dvc.yaml
