+++ mlflow_run_id.txt
@@ -2,8 +2,8 @@
 [INFO] Model directory: models/emotion_classification
 [INFO] Report directory: reports/emotion_classification
 [INFO] MLflow experiment: emotion_classification (ID: 7)
-  ⚠ GPU requested but not available, falling back to CPU
-[INFO] Using device: cpu
+  → Apple MPS (Metal Performance Shaders) available
+[INFO] Using device: mps
 [INFO] Image size: 48x48
 [INFO] Loading training data from data/processed/emotion_classification/train
 [INFO] Classes found: ['Angry', 'Disgust', 'Fear', 'Happy', 'Neutral', 'Sad', 'Surprise']
@@ -15,18 +15,18 @@
 [INFO] Starting MLflow run: emotion_classification_200_epochs
 [INFO] Logged training parameters to MLflow
 [INFO] Starting training for 1 epochs...
-Epoch 1/1 | Train Loss: 1.5099 | Train Acc: 0.4313 | Train Macro F1: 0.3341
-                 Val Acc: 0.4896 | Val Macro F1: 0.3781
-[INFO] ✓ New best model saved at epoch 1 (val_macro_f1=0.3781)
+Epoch 1/1 | Train Loss: 1.5132 | Train Acc: 0.4286 | Train Macro F1: 0.3321
+                 Val Acc: 0.4985 | Val Macro F1: 0.3840
+[INFO] ✓ New best model saved at epoch 1 (val_macro_f1=0.3840)
 [INFO] Training complete!
 [INFO] Loading best model from models/emotion_classification/best_model.pth
-[INFO] Saved test metrics to reports/emotion_classification/test_metrics_20251117_183410.json
-[INFO] Final validation accuracy: 0.4896
-[INFO] Final validation macro F1: 0.3781
+[INFO] Saved test metrics to reports/emotion_classification/test_metrics_20251118_013848.json
+[INFO] Final validation accuracy: 0.4985
+[INFO] Final validation macro F1: 0.3840
 [INFO] Uploading model and artifacts to MLflow...
 [INFO] ✓ Model uploaded to MLflow
 [INFO] ✓ Test metrics uploaded to MLflow
 [INFO] ✅ Training complete and all artifacts logged to MLflow!
-[INFO] Best model achieved val_macro_f1=0.3781 at epoch 1
-🏃 View run emotion_classification_200_epochs at: https://dagshub.com/pawarit.jamjod/MLOps_XOXO.mlflow/#/experiments/7/runs/5b33d4f286df4da4b2338101266ac0f5
+[INFO] Best model achieved val_macro_f1=0.3840 at epoch 1
+🏃 View run emotion_classification_200_epochs at: https://dagshub.com/pawarit.jamjod/MLOps_XOXO.mlflow/#/experiments/7/runs/391e2a0366b84b649de972aa6724b78d
 🧪 View experiment at: https://dagshub.com/pawarit.jamjod/MLOps_XOXO.mlflow/#/experiments/7
