+++ mlops_xoxo/age_gender/train.py
@@ -1,142 +1,266 @@
-#!/usr/bin/env python3
 """
-Entraînement GenderAge avec modèle PyTorch natif
+train.py - Train gender and age models with MLflow tracking
 """
-import torch, torch.nn as nn, torch.optim as optim
-from torch.utils.data import Dataset, DataLoader
-import cv2, numpy as np, pandas as pd
-from pathlib import Path
-from tqdm import tqdm
-import argparse
-import yaml
-with open("pipelines/age_gender/params.yaml", encoding="utf-8") as f:
-    params = yaml.safe_load(f)
 
-train_dir = params["dataset"]["processed_dir"] + "/train"
-val_dir = params["dataset"]["processed_dir"] + "/val"
-model_dir = "models/age_gender"
-
+import pandas as pd
 import torch
 import torch.nn as nn
+from torch.utils.data import Dataset, DataLoader
+from torchvision import transforms, models
+from PIL import Image
+from pathlib import Path
+import yaml
+from dotenv import load_dotenv
+import os
+from tqdm import tqdm
 
-class GenderAgeModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.features = nn.Sequential(
-            nn.Conv2d(3, 32, 3, 2, 1), nn.BatchNorm2d(32), nn.ReLU(inplace=True),
-            nn.Conv2d(32, 64, 3, 1, 1), nn.BatchNorm2d(64), nn.ReLU(inplace=True), nn.MaxPool2d(2),
-            nn.Conv2d(64, 128, 3, 1, 1), nn.BatchNorm2d(128), nn.ReLU(inplace=True), nn.MaxPool2d(2),
-            nn.Conv2d(128, 256, 3, 1, 1), nn.BatchNorm2d(256), nn.ReLU(inplace=True), nn.MaxPool2d(2),
-            nn.Conv2d(256, 512, 3, 1, 1), nn.BatchNorm2d(512), nn.ReLU(inplace=True), nn.AdaptiveAvgPool2d(1)
-        )
-        self.classifier = nn.Sequential(
-            nn.Dropout(0.5), nn.Linear(512, 128), nn.ReLU(inplace=True), nn.Dropout(0.3)
-        )
-        self.gender_head = nn.Linear(128, 2)
-        self.age_head = nn.Linear(128, 1)
+import mlflow
+import mlflow.pytorch
 
-    def forward(self, x):
-        x = self.features(x).view(x.size(0), -1)
-        x = self.classifier(x)
-        gender = self.gender_head(x)
-        age = self.age_head(x)
-        return torch.cat([gender, age], dim=1)
+# Load parameters
+with open('pipelines/age_gender/params.yaml', 'r') as f:
+    params = yaml.safe_load(f)
+
+# Configuration from params.yaml
+DATA_DIR = Path(params['dataset']['raw_dir'])
+TRAIN_FILE = Path(params['dataset']['processed_dir']) / "splits/train.csv"
+VAL_FILE = Path(params['dataset']['processed_dir']) / "splits/val.csv"
+OUTPUT_DIR = Path("models/age_gender")
+
+IMAGE_SIZE = params['preprocessing']['image_size']
+BATCH_SIZE = params['training']['batch_size']
+NUM_EPOCHS = params['training']['num_epochs']
+LEARNING_RATE = params['training']['learning_rate']
+SEED = params['seed']
 
 
-class GenderAgeLoss(nn.Module):
-    def __init__(self, gender_weight=1.0, age_weight=1.0):
-        super().__init__()
-        self.gender_loss = nn.CrossEntropyLoss()
-        self.age_loss = nn.MSELoss()
-        self.gender_weight = gender_weight
-        self.age_weight = age_weight
+load_dotenv()
+mlflow_uri = os.getenv("MLFLOW_TRACKING_URI")
+if mlflow_uri:
+    mlflow.set_tracking_uri(mlflow_uri)
+mlflow_username = os.getenv("MLFLOW_TRACKING_USERNAME")
+mlflow_password = os.getenv("MLFLOW_TRACKING_PASSWORD")
+if mlflow_username and mlflow_password:
+    os.environ["MLFLOW_TRACKING_USERNAME"] = mlflow_username
+    os.environ["MLFLOW_TRACKING_PASSWORD"] = mlflow_password
 
-    def forward(self, predictions, gender_target, age_target):
-        gender_pred = predictions[:, :2]
-        age_pred = predictions[:, 2]
-        loss_gender = self.gender_loss(gender_pred, gender_target)
-        loss_age = self.age_loss(age_pred, age_target)
-        total_loss = self.gender_weight * loss_gender + self.age_weight * loss_age
-        return total_loss, loss_gender, loss_age
+run_name = params['mlflow'].get('run_name', 'default_run')
+device_param = params['training']['device'].lower()
+if device_param == 'cuda' and torch.cuda.is_available():
+    DEVICE = torch.device('cuda')
+elif device_param == 'mps' and getattr(torch.backends, 'mps', None) is not None:
+    DEVICE = torch.device('mps')
+else:
+    DEVICE = torch.device('cpu')
+print("Using device:", DEVICE)
+
+# Set experiment
+mlflow.set_experiment(params['mlflow']['experiment_name'])
+
+
+class UTKFaceDataset(Dataset):
+    """Simple dataset loader"""
+    def __init__(self, df, data_dir, transform, task):
+        self.df = df.reset_index(drop=True)
+        self.data_dir = Path(data_dir)
+        self.transform = transform
+        self.task = task
     
-class GenderAgeDataset(Dataset):
-    def __init__(self, csv_path, images_dir, input_size=(112, 112)):
-        self.df = pd.read_csv(csv_path)
-        self.images_dir = Path(images_dir)
-        self.input_size = input_size
-    def __len__(self): return len(self.df)
+    def __len__(self):
+        return len(self.df)
+    
     def __getitem__(self, idx):
         row = self.df.iloc[idx]
-        img = cv2.imread(str(self.images_dir / row['image_name']))
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        img = cv2.resize(img, self.input_size).astype(np.float32) / 255.0
-        img = torch.from_numpy(img).permute(2, 0, 1)
-        gender = torch.tensor(row['gender'], dtype=torch.long)
-        age = torch.tensor(row['age'] / 100.0, dtype=torch.float32)
-        return img, gender, age
+        img = Image.open(self.data_dir / row['filename']).convert('RGB')
+        img = self.transform(img)
+        
+        if self.task == 'gender':
+            label = torch.tensor(row['gender'], dtype=torch.long)
+        else:
+            label = torch.tensor(row['age'], dtype=torch.float32)
+        
+        return img, label
 
 
-def train_epoch(model, loader, criterion, optimizer, device):
-    model.train(); total_loss, correct, total = 0, 0, 0
-    for imgs, genders, ages in tqdm(loader, desc="Training"):
-        imgs, genders, ages = imgs.to(device), genders.to(device), ages.to(device)
-        optimizer.zero_grad()
-        outputs = model(imgs)
-        loss, _, _ = criterion(outputs, genders, ages)
-        loss.backward(); optimizer.step()
-        total_loss += loss.item()
-        correct += (torch.argmax(outputs[:, :2], 1) == genders).sum().item()
-        total += genders.size(0)
-    return total_loss/len(loader), correct/total
+class SimpleModel(nn.Module):
+    """Lightweight MobileNetV2-based model"""
+    def __init__(self, num_outputs):
+        super().__init__()
+        self.backbone = models.mobilenet_v2(weights='IMAGENET1K_V1')
+        self.backbone.classifier = nn.Sequential(
+            nn.Dropout(p=params['model']['dropout']),
+            nn.Linear(self.backbone.last_channel, num_outputs)
+        )
+    
+    def forward(self, x):
+        out = self.backbone(x)
+        return out.squeeze() if out.shape[1] == 1 else out
 
 
-def validate(model, loader, criterion, device):
-    model.eval(); total_loss, correct, total, age_err = 0, 0, 0, []
-    with torch.no_grad():
-        for imgs, genders, ages in tqdm(loader, desc="Validation"):
-            imgs, genders, ages = imgs.to(device), genders.to(device), ages.to(device)
-            out = model(imgs)
-            loss, _, _ = criterion(out, genders, ages)
-            total_loss += loss.item()
-            correct += (torch.argmax(out[:, :2], 1) == genders).sum().item()
-            total += genders.size(0)
-            age_err.extend((out[:, 2]*100 - ages*100).abs().cpu().numpy())
-    return total_loss/len(loader), correct/total, np.mean(age_err)
+def get_transforms():
+    """Basic transforms"""
+    return transforms.Compose([
+        transforms.Resize((IMAGE_SIZE, IMAGE_SIZE)),
+        transforms.RandomHorizontalFlip(),
+        transforms.ToTensor(),
+        transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
+    ])
 
 
-def train_model(train_loader, val_loader, epochs, device, save_dir):
-    model = GenderAgeModel().to(device)
-    optimizer = optim.Adam(model.parameters(), lr=0.001)
-    scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, 'min', patience=5)
-    criterion = GenderAgeLoss(1.0, 2.0)
-    best_val = float('inf')
-    save_dir = Path(save_dir); save_dir.mkdir(exist_ok=True)
+def train_model(task, train_df, val_df):
+    """Train a single model with MLflow tracking"""
+    print(f"\n{'='*60}")
+    print(f"TRAINING: {task.upper()}")
+    print(f"{'='*60}")
+    
+    # Start MLflow run
+    with mlflow.start_run(run_name=f"{task}_training") as run:
+        
+        # Log parameters
+        mlflow.log_params({
+            "task": task,
+            "batch_size": BATCH_SIZE,
+            "num_epochs": NUM_EPOCHS,
+            "learning_rate": LEARNING_RATE,
+            "image_size": IMAGE_SIZE,
+            "backbone": params['model']['backbone'],
+            "dropout": params['model']['dropout'],
+            "device": str(DEVICE),
+            "seed": SEED
+        })
+        
+        # Log tags
+        mlflow.set_tags(params['mlflow']['tags'])
+        mlflow.set_tag("model_type", task)
+        
+        # Create datasets
+        train_dataset = UTKFaceDataset(train_df, DATA_DIR, get_transforms(), task)
+        val_dataset = UTKFaceDataset(val_df, DATA_DIR, get_transforms(), task)
+        
+        train_loader = DataLoader(train_dataset, batch_size=BATCH_SIZE, shuffle=True, num_workers=0)
+        val_loader = DataLoader(val_dataset, batch_size=BATCH_SIZE, shuffle=False, num_workers=0)
+        
+        # Create model
+        num_outputs = 2 if task == 'gender' else 1
+        model = SimpleModel(num_outputs).to(DEVICE)
+        
+        criterion = nn.CrossEntropyLoss() if task == 'gender' else nn.L1Loss()
+        optimizer = torch.optim.Adam(model.parameters(), lr=LEARNING_RATE)
+        
+                # Training loop
+        best_metric = 0 if task == 'gender' else float('inf')
 
-    for ep in range(1, epochs+1):
-        train_loss, train_acc = train_epoch(model, train_loader, criterion, optimizer, device)
-        val_loss, val_acc, val_mae = validate(model, val_loader, criterion, device)
-        scheduler.step(val_loss)
-        print(f"\nEpoch {ep}: train_loss={train_loss:.4f}, val_loss={val_loss:.4f}, val_acc={val_acc:.2%}, MAE={val_mae:.2f}")
-        if val_loss < best_val:
-            best_val = val_loss
-            torch.save(model.state_dict(), save_dir / 'best_genderage_native.pth')
-            print("✓ Nouveau meilleur modèle sauvegardé")
-    return model
+        for epoch in range(NUM_EPOCHS):
+            print(f"\nEpoch {epoch+1}/{NUM_EPOCHS}")
+
+            # ---------------------------
+            # TRAIN
+            # ---------------------------
+            model.train()
+            train_loss = 0
+
+            train_pbar = tqdm(train_loader, desc=f"[Train] Epoch {epoch+1}", leave=False)
+
+            for images, labels in train_pbar:
+                images, labels = images.to(DEVICE), labels.to(DEVICE)
+
+                optimizer.zero_grad()
+                outputs = model(images)
+                loss = criterion(outputs, labels)
+                loss.backward()
+                optimizer.step()
+
+                train_loss += loss.item()
+                train_pbar.set_postfix(loss=loss.item())
+
+            train_loss /= len(train_loader)
+
+            # ---------------------------
+            # VALIDATE
+            # ---------------------------
+            model.eval()
+            val_loss = 0
+            correct = 0
+            total = 0
+
+            val_pbar = tqdm(val_loader, desc=f"[Val] Epoch {epoch+1}", leave=False)
+
+            with torch.no_grad():
+                for images, labels in val_pbar:
+                    images, labels = images.to(DEVICE), labels.to(DEVICE)
+                    outputs = model(images)
+                    loss = criterion(outputs, labels)
+                    val_loss += loss.item()
+
+                    if task == 'gender':
+                        preds = torch.argmax(outputs, dim=1)
+                        correct += (preds == labels).sum().item()
+                    else:
+                        correct += torch.abs(outputs - labels).sum().item()
+
+                    total += labels.size(0)
+                    val_pbar.set_postfix(loss=loss.item())
+
+            val_loss /= len(val_loader)
+            metric = correct / total if task == 'gender' else correct / total
+
+            # Log metrics to MLflow
+            mlflow.log_metrics({
+                "train_loss": train_loss,
+                "val_loss": val_loss,
+                f"val_{'accuracy' if task == 'gender' else 'mae'}": metric
+            }, step=epoch)
+
+            print(f"Epoch {epoch+1}/{NUM_EPOCHS} - "
+                  f"TrainLoss: {train_loss:.4f} | ValLoss: {val_loss:.4f} | Metric: {metric:.4f}")
+
+            # Save best
+            is_best = (metric > best_metric) if task == 'gender' else (metric < best_metric)
+            if is_best:
+                best_metric = metric
+                print("  -> New best model found and saved!")
+        
+        # Log best metric
+        metric_name = "best_accuracy" if task == 'gender' else "best_mae"
+        mlflow.log_metric(metric_name, best_metric)
+        
+        # Save final model
+        OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
+        model_path = OUTPUT_DIR / f"{task}_model.pt"
+        torch.save(model.state_dict(), model_path)
+        print(f"\nSaved to: {model_path}")
+        
+        # Log model to MLflow
+        mlflow.pytorch.log_model(model, f"{task}_model")
+        
+        # Log model file as artifact
+        mlflow.log_artifact(str(model_path))
+        
+        print(f"✅ MLflow Run ID: {run.info.run_id}")
+
+
+def main():
+    print("=" * 60)
+    print(f"MODEL TRAINING (Device: {DEVICE})")
+    print("=" * 60)
+    
+    torch.manual_seed(SEED)
+    
+    # Load data
+    train_df = pd.read_csv(TRAIN_FILE)
+    val_df = pd.read_csv(VAL_FILE)
+    print(f"\nTrain: {len(train_df)}, Val: {len(val_df)}")
+    
+    # Train both models (separate MLflow runs)
+    train_model('gender', train_df, val_df)
+    train_model('age', train_df, val_df)
+    
+    print("\n" + "=" * 60)
+    print("TRAINING COMPLETED")
+    print("=" * 60)
+    
+    return 0
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--train-data', default=train_dir)
-    parser.add_argument('--val-data', default=val_dir)
-    parser.add_argument('--epochs', type=int, default=20)
-    parser.add_argument('--batch-size', type=int, default=32)
-    parser.add_argument('--device', default='cpu')
-    parser.add_argument('--output-dir', default=model_dir)
-    args = parser.parse_args()
-
-    device = args.device if torch.cuda.is_available() or args.device == "mps" else "cpu"
-    train_ds = GenderAgeDataset(Path(args.train_data)/'labels.csv', Path(args.train_data)/'images')
-    val_ds = GenderAgeDataset(Path(args.val_data)/'labels.csv', Path(args.val_data)/'images')
-    train_loader = DataLoader(train_ds, batch_size=args.batch_size, shuffle=True, num_workers=4)
-    val_loader = DataLoader(val_ds, batch_size=args.batch_size, shuffle=False, num_workers=4)
-    train_model(train_loader, val_loader, args.epochs, device, args.output_dir)
+    exit(main())
