+++ pipelines/age_gender/params.yaml
@@ -8,10 +8,13 @@ dataset:
     val: 0.15
 mlflow:
   experiment_id: "4"
-  experiment_name: face_embedding
+  experiment_name: age_gender_classification
   run_id: a90bc793e451424fa50694dee3032877
-  run_name: face_embedding_200_epochs
+  run_name: age_gender_classification_1_epochs
   tracking_uri: https://dagshub.com/pawarit.jamjod/MLOps_XOXO.mlflow
+  tags:
+    name: age_gender_classification
+    team: xoxo
 preprocessing:
   image_size: 160
   min_face_size: 40
@@ -19,10 +22,10 @@ seed: 42
 storage:
   dvc_remote: https://dagshub.com/pawarit.jamjod/MLOps_XOXO.dvc
 training:
-  batch_size: 64
+  batch_size: 16
+  num_epochs: 1
+  learning_rate: 0.001
   device: mps
-  embedding_size: 512
-  epochs: 200
-  lr: 0.0001
-  margin: 0.2
-  weight_decay: 0.0001
+model:
+  dropout: 0.5
+  backbone: mobilenet_v2
