+++ pipelines/emotion_classification/dvc.lock
@@ -5,9 +5,9 @@ stages:
     deps:
     - path: data/external/emotion_classification
       hash: md5
-      md5: ff0ef72ed24a2d15959c3b7d2a260356.dir
-      size: 160
-      nfiles: 1
+      md5: d751713988987e9331980363e24189ce.dir
+      size: 0
+      nfiles: 0
     - path: mlops_xoxo/emotion_classification/data_ingest.py
       hash: md5
       md5: dcf7274c91e69377dbeaebe1c7a773d5
@@ -15,14 +15,14 @@ stages:
     outs:
     - path: data/raw/emotion_classification/ingest_manifest.json
       hash: md5
-      md5: add0c91458aeba170040f9f6859f6d30
+      md5: 9ffabbf88798cec5c2464da464437746
       size: 155
   split:
     cmd: python mlops_xoxo/emotion_classification/data_split.py
     deps:
     - path: data/raw/emotion_classification
       hash: md5
-      md5: 25ebdfd4de058adfefa6efdc657ec4f3.dir
+      md5: 1d1e54eb0322b7cbc99161b4f0090ade.dir
       size: 62409633
       nfiles: 35891
     - path: mlops_xoxo/emotion_classification/data_split.py
@@ -40,7 +40,7 @@ stages:
     deps:
     - path: data/raw/emotion_classification
       hash: md5
-      md5: 25ebdfd4de058adfefa6efdc657ec4f3.dir
+      md5: 1d1e54eb0322b7cbc99161b4f0090ade.dir
       size: 62409633
       nfiles: 35891
     - path: mlops_xoxo/emotion_classification/data_validate.py
@@ -67,14 +67,14 @@ stages:
       size: 12813
     - path: pipelines/emotion_classification/params.yaml
       hash: md5
-      md5: a0e46ee4bd0e8cee16b230a16be15d6f
-      size: 673
+      md5: b2da20301c7b68b902d57de135f198a2
+      size: 671
     params:
       pipelines/emotion_classification/params.yaml:
         training:
           epochs: 1
           batch_size: 32
-          device: cuda
+          device: mps
           lr: 0.001
           fraction: 1.0
           early_stopping_patience: 4
@@ -82,13 +82,13 @@ stages:
     outs:
     - path: models/emotion_classification
       hash: md5
-      md5: 43deceacb03cf45dbdc76d63440801b9.dir
-      size: 178988368
+      md5: fb1a36a9640bc259f84737c617dfd076.dir
+      size: 178988560
       nfiles: 2
     - path: reports/emotion_classification/emissions.csv
       hash: md5
-      md5: 558d1755ce027eb79ad93fc31f695409
-      size: 772
+      md5: 8c37115999522dcce3280778eed96bd0
+      size: 787
   eval:
     cmd: python mlops_xoxo/emotion_classification/eval.py
     deps:
@@ -103,19 +103,19 @@ stages:
       size: 5306
     - path: models/emotion_classification
       hash: md5
-      md5: 43deceacb03cf45dbdc76d63440801b9.dir
-      size: 178988368
+      md5: fb1a36a9640bc259f84737c617dfd076.dir
+      size: 178988560
       nfiles: 2
     - path: pipelines/emotion_classification/params.yaml
       hash: md5
-      md5: a0e46ee4bd0e8cee16b230a16be15d6f
-      size: 673
+      md5: b2da20301c7b68b902d57de135f198a2
+      size: 671
     outs:
     - path: reports/emotion_classification/eval_metrics.json
       hash: md5
-      md5: 4b35588d6a1e6544dbec27c914029dc0
-      size: 1481
+      md5: 914820dda9bcb59aa3530d402859c835
+      size: 1487
     - path: reports/emotion_classification/eval_summary.txt
       hash: md5
-      md5: 8772a3c476e7451ab8d0defb32d1fe3b
+      md5: c2e1ad1fcca6af4eec74ca86f7f1362f
       size: 596
