+++ reports/age_gender/.gitignore
@@ -1,0 +1,2 @@
+/test_results
+/evaluation
