+++ reports/emotion_classification/.gitignore
@@ -1,0 +1,3 @@
+/eval_metrics.json
+/eval_summary.txt
+/emissions.csv
