+++ reports/emotion_classification/test_metrics_20251111_111906.json
@@ -1,0 +1,79 @@
+{
+  "accuracy": 0.5010224948875256,
+  "macro_f1": 0.3833824523158032,
+  "per_class_f1": [
+    0.23225806451612904,
+    0.0,
+    0.18181818181818182,
+    0.7251170046801872,
+    0.4698133918770582,
+    0.42554771630152244,
+    0.6491228070175439
+  ],
+  "confusion_matrix": [
+    [
+      108,
+      0,
+      23,
+      124,
+      124,
+      347,
+      16
+    ],
+    [
+      6,
+      0,
+      2,
+      22,
+      5,
+      45,
+      1
+    ],
+    [
+      34,
+      0,
+      91,
+      148,
+      87,
+      359,
+      49
+    ],
+    [
+      5,
+      0,
+      6,
+      1162,
+      57,
+      112,
+      6
+    ],
+    [
+      7,
+      0,
+      27,
+      156,
+      428,
+      301,
+      10
+    ],
+    [
+      16,
+      0,
+      34,
+      128,
+      149,
+      573,
+      11
+    ],
+    [
+      12,
+      0,
+      50,
+      117,
+      43,
+      45,
+      333
+    ]
+  ],
+  "report_text": "              precision    recall  f1-score   support\n\n       angry       0.57      0.15      0.23       742\n     disgust       0.00      0.00      0.00        81\n        fear       0.39      0.12      0.18       768\n       happy       0.63      0.86      0.73      1348\n     neutral       0.48      0.46      0.47       929\n         sad       0.32      0.63      0.43       911\n    surprise       0.78      0.56      0.65       600\n\n    accuracy                           0.50      5379\n   macro avg       0.45      0.40      0.38      5379\nweighted avg       0.52      0.50      0.47      5379\n"
+}
