+++ reports/emotion_classification/test_metrics_20251117_182550.json
@@ -1,0 +1,79 @@
+{
+  "accuracy": 0.4895682710183846,
+  "macro_f1": 0.3781395080936529,
+  "per_class_f1": [
+    0.3056092843326886,
+    0.0,
+    0.10361445783132531,
+    0.7211291702309667,
+    0.48467966573816157,
+    0.4285085789705235,
+    0.6034353995519044
+  ],
+  "confusion_matrix": [
+    [
+      158,
+      0,
+      31,
+      60,
+      115,
+      249,
+      59
+    ],
+    [
+      27,
+      0,
+      5,
+      4,
+      9,
+      22,
+      6
+    ],
+    [
+      59,
+      0,
+      43,
+      54,
+      107,
+      244,
+      186
+    ],
+    [
+      48,
+      0,
+      9,
+      843,
+      85,
+      174,
+      55
+    ],
+    [
+      17,
+      0,
+      12,
+      64,
+      435,
+      263,
+      47
+    ],
+    [
+      48,
+      0,
+      26,
+      43,
+      165,
+      487,
+      44
+    ],
+    [
+      5,
+      0,
+      11,
+      56,
+      41,
+      21,
+      404
+    ]
+  ],
+  "report_text": "              precision    recall  f1-score   support\n\n       Angry       0.44      0.24      0.31       672\n     Disgust       0.00      0.00      0.00        73\n        Fear       0.31      0.06      0.10       693\n       Happy       0.75      0.69      0.72      1214\n     Neutral       0.45      0.52      0.48       838\n         Sad       0.33      0.60      0.43       813\n    Surprise       0.50      0.75      0.60       538\n\n    accuracy                           0.49      4841\n   macro avg       0.40      0.41      0.38      4841\nweighted avg       0.48      0.49      0.46      4841\n"
+}
