+++ reports/emotion_classification/test_metrics_20251118_013600.json
@@ -1,0 +1,79 @@
+{
+  "accuracy": 0.4984507333195621,
+  "macro_f1": 0.38399891473376613,
+  "per_class_f1": [
+    0.37850467289719625,
+    0.0,
+    0.10778443113772455,
+    0.7123188405797102,
+    0.4835988753514527,
+    0.3818301514154049,
+    0.6239554317548747
+  ],
+  "confusion_matrix": [
+    [
+      243,
+      0,
+      25,
+      91,
+      173,
+      116,
+      24
+    ],
+    [
+      38,
+      0,
+      6,
+      6,
+      11,
+      11,
+      1
+    ],
+    [
+      103,
+      0,
+      45,
+      125,
+      171,
+      134,
+      115
+    ],
+    [
+      58,
+      0,
+      9,
+      983,
+      84,
+      50,
+      30
+    ],
+    [
+      50,
+      0,
+      20,
+      145,
+      516,
+      94,
+      13
+    ],
+    [
+      100,
+      0,
+      26,
+      120,
+      257,
+      290,
+      20
+    ],
+    [
+      20,
+      0,
+      11,
+      76,
+      84,
+      11,
+      336
+    ]
+  ],
+  "report_text": "              precision    recall  f1-score   support\n\n       Angry       0.40      0.36      0.38       672\n     Disgust       0.00      0.00      0.00        73\n        Fear       0.32      0.06      0.11       693\n       Happy       0.64      0.81      0.71      1214\n     Neutral       0.40      0.62      0.48       838\n         Sad       0.41      0.36      0.38       813\n    Surprise       0.62      0.62      0.62       538\n\n    accuracy                           0.50      4841\n   macro avg       0.40      0.40      0.38      4841\nweighted avg       0.47      0.50      0.46      4841\n"
+}
